+++ src/main/java/nextstep/presentation/ReservationController.java
@@ -3,8 +3,8 @@ package nextstep.presentation;
 import java.net.URI;
 import java.util.List;
 import nextstep.application.ReservationService;
-import nextstep.presentation.dto.ReservationRequest;
-import nextstep.presentation.dto.ReservationResponse;
+import nextstep.presentation.dto.reservation.ReservationRequest;
+import nextstep.presentation.dto.reservation.ReservationResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
