+++ src/main/java/nextstep/presentation/ScheduleController.java
@@ -3,8 +3,8 @@ package nextstep.presentation;
 import java.net.URI;
 import java.util.List;
 import nextstep.application.ScheduleService;
-import nextstep.presentation.dto.ScheduleRequest;
-import nextstep.presentation.dto.ScheduleResponse;
+import nextstep.presentation.dto.schedule.ScheduleRequest;
+import nextstep.presentation.dto.schedule.ScheduleResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
