+++ src/main/java/nextstep/presentation/ThemeController.java
@@ -3,8 +3,8 @@ package nextstep.presentation;
 import java.net.URI;
 import java.util.List;
 import nextstep.application.ThemeService;
-import nextstep.presentation.dto.ThemeRequest;
-import nextstep.presentation.dto.ThemeResponse;
+import nextstep.presentation.dto.theme.ThemeRequest;
+import nextstep.presentation.dto.theme.ThemeResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
