+++ src/main/java/nextstep/presentation/dto/auth/AuthResponse.java
@@ -1,0 +1,17 @@
+package nextstep.presentation.dto.auth;
+
+public class AuthResponse {
+
+    private String accessToken;
+
+    private AuthResponse() {
+    }
+
+    public AuthResponse(String accessToken) {
+        this.accessToken = accessToken;
+    }
+
+    public String getAccessToken() {
+        return accessToken;
+    }
+}
