+++ src/main/java/nextstep/presentation/dto/member/MemberRequest.java
@@ -1,4 +1,4 @@
-package nextstep.presentation.dto;
+package nextstep.presentation.dto.member;
 
 import nextstep.domain.Member;
