+++ src/main/java/nextstep/presentation/dto/reservation/ReservationResponse.java
@@ -1,4 +1,4 @@
-package nextstep.presentation.dto;
+package nextstep.presentation.dto.reservation;
 
 public class ReservationResponse {
