+++ src/main/java/nextstep/presentation/dto/schedule/ScheduleRequest.java
@@ -1,4 +1,4 @@
-package nextstep.presentation.dto;
+package nextstep.presentation.dto.schedule;
 
 public class ScheduleRequest {
