+++ src/main/java/nextstep/presentation/dto/theme/ThemeResponse.java
@@ -1,4 +1,4 @@
-package nextstep.presentation.dto;
+package nextstep.presentation.dto.theme;
 
 public class ThemeResponse {
