+++ src/test/java/nextstep/integration/ScheduleServiceTest.java
@@ -9,10 +9,10 @@ import nextstep.application.ReservationService;
 import nextstep.application.ScheduleService;
 import nextstep.application.ThemeService;
 import nextstep.exception.ScheduleException;
-import nextstep.presentation.dto.ScheduleRequest;
-import nextstep.presentation.dto.ScheduleResponse;
-import nextstep.presentation.dto.ThemeRequest;
-import nextstep.presentation.dto.ThemeResponse;
+import nextstep.presentation.dto.schedule.ScheduleRequest;
+import nextstep.presentation.dto.schedule.ScheduleResponse;
+import nextstep.presentation.dto.theme.ThemeRequest;
+import nextstep.presentation.dto.theme.ThemeResponse;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
