+++ src/test/java/nextstep/integration/ThemeServiceTest.java
@@ -7,8 +7,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.List;
 import nextstep.application.ThemeService;
 import nextstep.exception.ThemeException;
-import nextstep.presentation.dto.ThemeRequest;
-import nextstep.presentation.dto.ThemeResponse;
+import nextstep.presentation.dto.theme.ThemeRequest;
+import nextstep.presentation.dto.theme.ThemeResponse;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
